+++ examples/TypeScriptMessaging/index.js
@@ -1,5 +1,4 @@
 import 'react-native-gesture-handler';
-import 'react-native-get-random-values';
 import { AppRegistry } from 'react-native';
 
 import App from './App';