+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   - [react-native-reanimated](https://docs.swmansion.com/react-native-reanimated/docs/installation)
   - [react-native-svg](https://github.com/react-native-svg/react-native-svg#installation)
 
+- Removed peer dependencies for:
+  - [react-native-get-random-values](https://github.com/LinusU/react-native-get-random-values)
+
 - Removed styled-components:
   - Removed dot notation for theming applications
   - Removed css string notation for styles on theme